Cleanup first (typos, deprecated entities...)
